Switch auto field stats to an item pipeline
