+++ scrapy_zyte_api/poet.py
@@ -1,0 +1,1 @@
+from ._poet_item_pipelines import ScrapyZyteAPIPoetItemPipeline
