+++ setup.cfg
@@ -4,5 +4,7 @@ max-line-length = 88
 max-complexity = 18
 select = B,C,E,F,W,T4
 per-file-ignores =
+    tests/test_auto_field_stats.py: E402
     tests/test_providers.py: E402
     scrapy_zyte_api/__init__.py: F401
+    scrapy_zyte_api/poet.py: F401
