+++ setup.py
@@ -33,7 +33,8 @@ setuptools.setup(
             "andi>=0.6.0",
             "scrapy-poet>=0.22.3",
             "web-poet>=0.17.0",
-            "zyte-common-items>=0.20.0",
+            # "zyte-common-items>=0.20.0",
+            "zyte-common-items @ git+https://github.com/Gallaecio/zyte-common-items.git@auto-fields",
         ]
     },
     classifiers=[
